% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_metrics.R
\name{availability}
\alias{availability}
\title{Availability calculation}
\usage{
availability(mtbf, mttr)
}
\arguments{
\item{mtbf}{Mean time between failures}

\item{mttr}{Mean time to repair}
}
\value{
Numeric, availability (between 0~1)
}
\description{
Calculates system availability: A = MTBF / (MTBF + MTTR)
}
\examples{
availability(mtbf = 1000, mttr = 10)   # ~ 0.99
}
