% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{calc_power}
\alias{calc_power}
\title{Power calculation}
\usage{
calc_power(
  n,
  effect_size,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  test_type = c("t", "z")
)
}
\arguments{
\item{n}{Sample size}

\item{effect_size}{Cohen's d effect size}

\item{alpha}{Significance level (default 0.05)}

\item{alternative}{Alternative hypothesis direction}

\item{test_type}{Test type ("t" or "z")}
}
\value{
Power (value between 0 and 1)
}
\description{
Given sample size, effect size, and significance level, calculates power.
}
\examples{
calc_power(n = 50, effect_size = 0.5, alpha = 0.05)
}
