% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fit.R
\name{calc_qq_data}
\alias{calc_qq_data}
\title{QQ Data Calculation}
\usage{
calc_qq_data(x, dist = "norm", params = list())
}
\arguments{
\item{x}{Numeric vector}

\item{dist}{Distribution name}

\item{params}{Distribution parameter list}
}
\value{
Data frame containing theoretical (theoretical quantiles), sample (sample quantiles)
}
\description{
Calculate theoretical quantiles and sample quantiles required for QQ plot.
}
\examples{
set.seed(123)
x <- rnorm(100)
qq_data <- calc_qq_data(x, "norm", list(mean = 0, sd = 1))
}
