% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_analyze}
\alias{desc_analyze}
\title{Batch descriptive statistics}
\usage{
desc_analyze(data, vars = NULL, conf = 0.95)
}
\arguments{
\item{data}{Data frame}

\item{vars}{Variable vector}

\item{conf}{Confidence level}
}
\value{
Statistics result list
}
\description{
Batch descriptive statistics
}
