% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_box}
\alias{desc_box}
\title{Descriptive statistics boxplot}
\usage{
desc_box(s, theme = "prism")
}
\arguments{
\item{s}{Result returned by desc_calc}

\item{theme}{Theme (default "prism")}
}
\value{
ggplot object
}
\description{
Descriptive statistics boxplot
}
