% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_box_with_stats}
\alias{desc_box_with_stats}
\title{Descriptive statistics boxplot (with statistics table)}
\usage{
desc_box_with_stats(s, theme = "prism")
}
\arguments{
\item{s}{Result returned by desc_calc}

\item{theme}{Theme (default "prism")}
}
\value{
ggplot object
}
\description{
Descriptive statistics boxplot (with statistics table)
}
