% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_calc}
\alias{desc_calc}
\title{Descriptive statistics calculation}
\usage{
desc_calc(x, conf = 0.95)
}
\arguments{
\item{x}{Numeric vector}

\item{conf}{Confidence level}
}
\value{
Descriptive statistics list
}
\description{
Descriptive statistics calculation
}
