% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_plot}
\alias{desc_plot}
\title{Descriptive statistics combined plot}
\usage{
desc_plot(s, theme = "academic")
}
\arguments{
\item{s}{Result returned by desc_calc}

\item{theme}{Theme (default "academic")}
}
\value{
patchwork object
}
\description{
Descriptive statistics combined plot
}
