% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_stats_table}
\alias{desc_stats_table}
\title{Statistics table plot}
\usage{
desc_stats_table(s, theme = "prism")
}
\arguments{
\item{s}{Result returned by desc_calc}

\item{theme}{Theme (default "prism")}
}
\value{
ggplot object
}
\description{
Statistics table plot
}
