% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_summary_table}
\alias{desc_summary_table}
\title{Descriptive statistics summary table}
\usage{
desc_summary_table(results)
}
\arguments{
\item{results}{Result returned by desc_analyze}
}
\value{
data.frame
}
\description{
Descriptive statistics summary table
}
