% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_to_excel}
\alias{desc_to_excel}
\title{Export to Excel report}
\usage{
desc_to_excel(results, path = NULL, excel_exporter = NULL)
}
\arguments{
\item{results}{Result returned by desc_analyze}

\item{path}{Output path}

\item{excel_exporter}{ExcelExporter instance (from iQualityR.core)}
}
\description{
Export to Excel report
}
