% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{effect_size}
\alias{effect_size}
\title{Effect size calculation}
\usage{
effect_size(
  type = c("cohens_d", "hedges_g", "eta_squared", "cohens_h", "r"),
  ...
)
}
\arguments{
\item{type}{Effect size type ("cohens_d", "hedges_g", "eta_squared", "cohens_h", "r")}

\item{...}{Different parameters based on type}
}
\value{
Effect size value
}
\description{
Calculates common effect size metrics (Cohen's d, Hedges' g, eta squared, etc.).
}
\examples{
effect_size("cohens_d", mean1 = 10, mean2 = 11, sd_pooled = 1.5)
}
