% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fit.R
\name{empirical_distribution}
\alias{empirical_distribution}
\title{Empirical Distribution Function}
\usage{
empirical_distribution(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
List containing ecdf (empirical distribution function object),
points (points for plotting), n (sample size)
}
\description{
Calculate and return the empirical distribution function of the data.
}
\examples{
set.seed(123)
x <- rnorm(100)
edf <- empirical_distribution(x)
edf$ecdf(0)
}
