% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{get_A3}
\alias{get_A3}
\title{Calculate A3 Constant}
\usage{
get_A3(n, k = 3)
}
\arguments{
\item{n}{Sample size(s).}

\item{k}{Sigma level (default 3).}
}
\description{
\strong{A3 = k / (c4 * sqrt(n))}
}
\details{
Used in X-bar charts when variation is estimated via average standard deviation (s-bar).
}
