% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{get_B3}
\alias{get_B3}
\title{Calculate B3 Constant}
\usage{
get_B3(n, k = 3)
}
\arguments{
\item{n}{Sample size(s).}

\item{k}{Sigma level (default 3).}
}
\description{
\strong{B3 = max(0, 1 - (k/c4) * sqrt(1 - c4^2))}
}
\details{
Used to determine the lower control limit for S-charts: \strong{LCL = B3 * s-bar}.
}
