% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{get_B4}
\alias{get_B4}
\title{Calculate B3 and B4 Constants}
\usage{
get_B4(n, k = 3)
}
\arguments{
\item{n}{Sample size(s).}

\item{k}{Sigma level (default 3).}
}
\description{
\strong{B4 = 1 + (k/c4) * sqrt(1 - c4^2)}, \strong{B3 = max(0, 1 - (k/c4) * sqrt(1 - c4^2))}
}
\details{
\strong{1. Purpose:} Used for S-charts: \strong{UCL = B4 * s-bar} and \strong{LCL = B3 * s-bar}.
}
