% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{get_c4}
\alias{get_c4}
\title{Calculate c4 Constant}
\usage{
get_c4(n)
}
\arguments{
\item{n}{Sample size(s).}
}
\value{
The value of c4 constant.
}
\description{
Calculate c4 Constant
}
\details{
\strong{1. Source and Theoretical Logic:} The sample standard deviation s is the square root of an unbiased estimator of variance sigma^2. According to Jensen's Inequality, E(s) < sigma. The c4 factor corrects this bias so that \strong{E(s/c4) = sigma}.

\strong{2. Mathematical Formula:} Derived using the properties of the Chi-distribution:
\strong{c4(n) = sqrt(2/(n-1)) * Gamma(n/2) / Gamma((n-1)/2)}

\strong{3. References:} #' * Shewhart, W. A. (1931). "Economic Control of Quality of Manufactured Product".
}
\examples{
get_c4(5) # Returns approximately 0.9400
}
