% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{get_c4_prime}
\alias{get_c4_prime}
\title{Calculate c4_prime Constant}
\usage{
get_c4_prime(n, B = 1e+06)
}
\arguments{
\item{n}{Sample size(s).}

\item{B}{Iterations for simulation (default 1e6).}
}
\value{
The value of c4_prime constant.
}
\description{
Calculate c4_prime Constant
}
\details{
c4_prime is used for bias correction when sigma is estimated via
Mean Successive Squared Differences (MSSD).
}
