% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_registry.R
\name{get_dist_info}
\alias{get_dist_info}
\title{Get detailed distribution information}
\usage{
get_dist_info(dist_type)
}
\arguments{
\item{dist_type}{Distribution type name}
}
\value{
Distribution information list
}
\description{
Returns complete metadata for specified distribution
}
\examples{
get_dist_info("norm")
}
