% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_prob.R
\name{get_prob_dist_info}
\alias{get_prob_dist_info}
\title{Get distribution information}
\usage{
get_prob_dist_info(type)
}
\arguments{
\item{type}{Distribution type}
}
\value{
Distribution information list
}
\description{
Returns detailed information for specified distribution.
}
\examples{
get_prob_dist_info("norm")
}
