% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTestPlotter.R
\name{HTestPlotter}
\alias{HTestPlotter}
\title{HTestPlotter: Hypothesis Test Plotting Coordinator}
\description{
Automatically select appropriate visualization based on test results.
All plotting functions reuse iQualityR.plot package functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{theme_obj}}{Theme object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HTestPlotter-new}{\code{HTestPlotter$new()}}
\item \href{#method-HTestPlotter-plot}{\code{HTestPlotter$plot()}}
\item \href{#method-HTestPlotter-plot_curve}{\code{HTestPlotter$plot_curve()}}
\item \href{#method-HTestPlotter-plot_box}{\code{HTestPlotter$plot_box()}}
\item \href{#method-HTestPlotter-plot_combined}{\code{HTestPlotter$plot_combined()}}
\item \href{#method-HTestPlotter-set_theme}{\code{HTestPlotter$set_theme()}}
\item \href{#method-HTestPlotter-clone}{\code{HTestPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$new(theme = "academic")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme name or IqrTheme object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
Automatically plot based on test results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$plot(
  result,
  x = NULL,
  y = NULL,
  plot_type = "auto",
  show_table = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result list (HTestAnalyzer output)}

\item{\code{x}}{Sample data (required for box plot)}

\item{\code{y}}{Second sample (required for two-sample tests)}

\item{\code{plot_type}}{Plot type ("auto", "curve", "box", "combined")}

\item{\code{show_table}}{Whether to display statistics table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 or patchwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-plot_curve"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-plot_curve}{}}}
\subsection{Method \code{plot_curve()}}{
Plot rejection region
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$plot_curve(result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-plot_box"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-plot_box}{}}}
\subsection{Method \code{plot_box()}}{
Plot boxplot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$plot_box(result, x = NULL, y = NULL, show_table = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}

\item{\code{x}}{Sample data}

\item{\code{y}}{Second sample}

\item{\code{show_table}}{Whether to display statistics table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-plot_combined"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-plot_combined}{}}}
\subsection{Method \code{plot_combined()}}{
Plot combined figure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$plot_combined(result, x = NULL, y = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}

\item{\code{x}}{Sample data}

\item{\code{y}}{Second sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
patchwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-set_theme"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-set_theme}{}}}
\subsection{Method \code{set_theme()}}{
Set theme
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$set_theme(theme)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme name or IqrTheme object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HTestPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
