% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{iqr_desc}
\alias{iqr_desc}
\title{Descriptive statistics quick analysis function}
\usage{
iqr_desc(
  data,
  vars = NULL,
  conf = 0.95,
  plot = FALSE,
  report = FALSE,
  theme = "prism"
)
}
\arguments{
\item{data}{Data frame or numeric vector}

\item{vars}{Variable vector}

\item{conf}{Confidence level}

\item{plot}{Whether to plot}

\item{report}{Whether to generate report}

\item{theme}{Theme}
}
\value{
Invisibly returns results list, or outputs plots and report simultaneously
}
\description{
Descriptive statistics quick analysis function
}
