% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_regression.R
\name{iqr_regression}
\alias{iqr_regression}
\title{iqr_regression: Regression entry class}
\description{
L3 integrator for regression.
Supported: lm_fit, logit_fit, poisson_fit, cox_fit, pls_fit, stepwise_fit,
best_subset_fit, mars_fit, spline_fit.
}
\examples{
set.seed(123)
x <- rnorm(30); y <- 2 + 3 * x + rnorm(30, sd = 0.5)
df <- data.frame(y = y, x = x)
reg <- iqr_regression$new()
reg$run("lm_fit", formula = y ~ x, data = df)
reg$interpret(audience = "manager")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-iqr_regression-new}{\code{iqr_regression$new()}}
\item \href{#method-iqr_regression-run}{\code{iqr_regression$run()}}
\item \href{#method-iqr_regression-plot}{\code{iqr_regression$plot()}}
\item \href{#method-iqr_regression-interpret}{\code{iqr_regression$interpret()}}
\item \href{#method-iqr_regression-report}{\code{iqr_regression$report()}}
\item \href{#method-iqr_regression-analyze}{\code{iqr_regression$analyze()}}
\item \href{#method-iqr_regression-set_theme}{\code{iqr_regression$set_theme()}}
\item \href{#method-iqr_regression-clone}{\code{iqr_regression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-new"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$new(theme = "academic")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-run"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$run(
  model_type,
  ...,
  plot = FALSE,
  plot_type = "auto",
  interpret = FALSE,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-plot"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$plot(plot_type = "auto", show_table = FALSE, theme_obj = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-interpret"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-interpret}{}}}
\subsection{Method \code{interpret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$interpret(audience = "manager")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-report"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$report(
  format = c("data.frame", "console", "excel"),
  path = NULL,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-analyze}{}}}
\subsection{Method \code{analyze()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$analyze(
  model_type,
  ...,
  plot_type = "auto",
  audience = "manager"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-set_theme"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-set_theme}{}}}
\subsection{Method \code{set_theme()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$set_theme(theme_style = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iqr_regression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-iqr_regression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{iqr_regression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
