% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_registry.R
\name{list_available_dists}
\alias{list_available_dists}
\title{List all available distributions}
\usage{
list_available_dists()
}
\value{
Data frame containing type, description, is_discrete, support columns
}
\description{
Returns names and descriptions of all registered distributions
}
\examples{
list_available_dists()
}
