% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_prob.R
\name{list_prob_distributions}
\alias{list_prob_distributions}
\title{List all available probability distributions}
\usage{
list_prob_distributions()
}
\value{
Data frame
}
\description{
Returns names and descriptions of all registered distributions.
}
\examples{
list_prob_distributions()
}
