% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_rules.R
\name{list_spc_rules}
\alias{list_spc_rules}
\title{List available SPC rules}
\usage{
list_spc_rules()
}
\value{
A data.frame with rule numbers and descriptions.
}
\description{
Returns a list of all available SPC out-of-control rules with descriptions.
}
\examples{
list_spc_rules()
}
