% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_prob.R
\name{prob_plot}
\alias{prob_plot}
\title{Convenience probability plotting function}
\usage{
prob_plot(
  type,
  params = list(),
  x = NULL,
  calc_type = "lower",
  show_cdf = TRUE,
  facet = FALSE,
  loc = 0,
  ...
)
}
\arguments{
\item{type}{Distribution type}

\item{params}{Distribution parameter list}

\item{x}{Input value (for annotation)}

\item{calc_type}{Calculation type}

\item{show_cdf}{Whether to show CDF plot}

\item{facet}{Whether to display as facets}

\item{loc}{Location offset}

\item{...}{Other parameters}
}
\value{
ggplot/patchwork object
}
\description{
Directly plot probability distribution without creating R6 object.
}
\examples{
prob_plot(type = "norm", params = list(mean = 0, sd = 1),
          x = 1.96, calc_type = "upper", show_cdf = TRUE)
}
