% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbPlotter.R
\name{ProbPlotter}
\alias{ProbPlotter}
\title{ProbPlotter: Probability distribution plotting engine}
\description{
Responsible for generating PDF/CDF plots, shaded areas, label annotations, etc.
Uses ggrepel to avoid label overlap, reuses IqrTheme for consistent visual style.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProbPlotter-render}{\code{ProbPlotter$render()}}
\item \href{#method-ProbPlotter-clone}{\code{ProbPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbPlotter-render"></a>}}
\if{latex}{\out{\hypertarget{method-ProbPlotter-render}{}}}
\subsection{Method \code{render()}}{
Render plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbPlotter$render(nodes, calc_results, facet, show_cdf, mode, theme_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodes}}{Distribution node list}

\item{\code{calc_results}}{Calculation result list}

\item{\code{facet}}{Whether to display as facets}

\item{\code{show_cdf}}{Whether to show CDF plot}

\item{\code{mode}}{Calculation mode}

\item{\code{theme_obj}}{IqrTheme object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot/patchwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProbPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
