% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbReporter.R
\name{ProbReporter}
\alias{ProbReporter}
\title{ProbReporter: Probability distribution report engine}
\description{
Responsible for outputting calculation results in table and text format, supports:
\itemize{
\item Console printing
\item Data frame export
\item Excel report export
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ProbReporter-print_console}{\code{ProbReporter$print_console()}}
\item \href{#method-ProbReporter-to_dataframe}{\code{ProbReporter$to_dataframe()}}
\item \href{#method-ProbReporter-export_excel}{\code{ProbReporter$export_excel()}}
\item \href{#method-ProbReporter-clone}{\code{ProbReporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbReporter-print_console"></a>}}
\if{latex}{\out{\hypertarget{method-ProbReporter-print_console}{}}}
\subsection{Method \code{print_console()}}{
Print report to console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbReporter$print_console(calc_results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{calc_results}}{Calculation result list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbReporter-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-ProbReporter-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Export as data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbReporter$to_dataframe(calc_results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{calc_results}}{Calculation result list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbReporter-export_excel"></a>}}
\if{latex}{\out{\hypertarget{method-ProbReporter-export_excel}{}}}
\subsection{Method \code{export_excel()}}{
Export to Excel report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbReporter$export_excel(calc_results, nodes, path, excel_exporter)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{calc_results}}{Calculation result list}

\item{\code{nodes}}{Distribution node list}

\item{\code{path}}{Output path}

\item{\code{excel_exporter}}{ExcelExporter instance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProbReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProbReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProbReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
