% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_registry.R
\name{register_dist}
\alias{register_dist}
\title{Register New Distribution}
\usage{
register_dist(
  dist_type,
  d_func,
  p_func,
  q_func,
  defaults,
  validate = NULL,
  is_discrete = FALSE,
  support = "Custom",
  description = "User-defined distribution"
)
}
\arguments{
\item{dist_type}{Distribution type name}

\item{d_func}{Probability density/mass function}

\item{p_func}{Cumulative distribution function}

\item{q_func}{Quantile function}

\item{defaults}{Default parameter values list}

\item{validate}{Parameter validation function}

\item{is_discrete}{Whether it is a discrete distribution}

\item{support}{Support description}

\item{description}{Distribution description}
}
\value{
Invisible NULL
}
\description{
Add a new distribution to the registry
}
\examples{
# Register a triangular-shaped custom distribution (then unregister it)
my_d <- function(x, p) pmax(0, 1 - abs(x - p$center) / p$half_width)
my_p <- function(q, p) pmax(0, pmin(1, 0.5 + (q - p$center) / p$half_width))
my_q <- function(p_vec, p) p$center + (p_vec - 0.5) * p$half_width
register_dist("my_tri",
              d_func = my_d, p_func = my_p, q_func = my_q,
              defaults = list(center = 0, half_width = 1),
              is_discrete = FALSE, support = "R",
              description = "Custom triangular distribution")
list_available_dists()
unregister_dist("my_tri")
}
