% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iqr_regression.R
\name{regression_interpret}
\alias{regression_interpret}
\title{Convenience regression interpretation function}
\usage{
regression_interpret(result, audience = "manager")
}
\arguments{
\item{result}{A stat_result from \code{\link[=regression_run]{regression_run()}}.}

\item{audience}{Audience level.}
}
\value{
Interpretation string (invisibly).
}
\description{
Convenience regression interpretation function
}
