% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionPlotter.R
\name{RegressionPlotter}
\alias{RegressionPlotter}
\title{RegressionPlotter: Regression Visualization}
\description{
L2 presentation engine for regression models.
$plot(result, plot_type = "auto", show_table = FALSE, theme_obj = NULL)

Plot types: "residual", "coef", "subset" (best_subset_fit),
"roc", "pr", "lift" (logit_fit classification curves),
"spline" (spline_fit / mars_fit data + fitted curve).
}
\section{Super class}{
\code{\link[iQualityR.stat:StatPlotter]{iQualityR.stat::StatPlotter}} -> \code{RegressionPlotter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegressionPlotter-new}{\code{RegressionPlotter$new()}}
\item \href{#method-RegressionPlotter-plot}{\code{RegressionPlotter$plot()}}
\item \href{#method-RegressionPlotter-clone}{\code{RegressionPlotter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.stat" data-topic="StatPlotter" data-id="set_theme"><a href='../../iQualityR.stat/html/StatPlotter.html#method-StatPlotter-set_theme'><code>iQualityR.stat::StatPlotter$set_theme()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionPlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionPlotter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionPlotter$new(theme = "academic")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionPlotter$plot(
  result,
  plot_type = "auto",
  show_table = FALSE,
  theme_obj = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
