% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionReporter.R
\name{RegressionReporter}
\alias{RegressionReporter}
\title{RegressionReporter: Regression Report Output}
\description{
Convert a regression stat_result into console / data.frame / excel.
$report(result, format = c("data.frame","console","excel"), path = NULL, audience = "manager")
}
\section{Super classes}{
\code{\link[iQualityR.core:IqrReporter]{iQualityR.core::IqrReporter}} -> \code{\link[iQualityR.stat:StatReporter]{iQualityR.stat::StatReporter}} -> \code{RegressionReporter}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RegressionReporter-new}{\code{RegressionReporter$new()}}
\item \href{#method-RegressionReporter-report}{\code{RegressionReporter$report()}}
\item \href{#method-RegressionReporter-print_console}{\code{RegressionReporter$print_console()}}
\item \href{#method-RegressionReporter-to_dataframe}{\code{RegressionReporter$to_dataframe()}}
\item \href{#method-RegressionReporter-export_excel}{\code{RegressionReporter$export_excel()}}
\item \href{#method-RegressionReporter-clone}{\code{RegressionReporter$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrReporter" data-id="export"><a href='../../iQualityR.core/html/IqrReporter.html#method-IqrReporter-export'><code>iQualityR.core::IqrReporter$export()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrReporter" data-id="export_rmd"><a href='../../iQualityR.core/html/IqrReporter.html#method-IqrReporter-export_rmd'><code>iQualityR.core::IqrReporter$export_rmd()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrReporter" data-id="register"><a href='../../iQualityR.core/html/IqrReporter.html#method-IqrReporter-register'><code>iQualityR.core::IqrReporter$register()</code></a></span></li>
<li><span class="pkg-link" data-pkg="iQualityR.core" data-topic="IqrReporter" data-id="render"><a href='../../iQualityR.core/html/IqrReporter.html#method-IqrReporter-render'><code>iQualityR.core::IqrReporter$render()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$new(theme = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-report"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-report}{}}}
\subsection{Method \code{report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$report(
  result,
  format = c("data.frame", "console", "excel"),
  path = NULL,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-print_console"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-print_console}{}}}
\subsection{Method \code{print_console()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$print_console(
  result,
  interpret = TRUE,
  interpreter = NULL,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$to_dataframe(result)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-export_excel"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-export_excel}{}}}
\subsection{Method \code{export_excel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$export_excel(result, path = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RegressionReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RegressionReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RegressionReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
