% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_registry.R
\name{unregister_dist}
\alias{unregister_dist}
\title{Unregister Distribution}
\usage{
unregister_dist(dist_type)
}
\arguments{
\item{dist_type}{Distribution type name}
}
\value{
Invisible NULL
}
\description{
Remove a specified distribution from the registry
}
