% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_hist}
\alias{desc_hist}
\title{Descriptive statistics histogram}
\usage{
desc_hist(s, theme = "academic")
}
\arguments{
\item{s}{Result returned by desc_calc}

\item{theme}{Theme (default "academic")}
}
\value{
ggplot object
}
\description{
Descriptive statistics histogram
}
