% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTestReporter.R
\name{HTestReporter}
\alias{HTestReporter}
\title{HTestReporter: Hypothesis Test Report Output}
\description{
Output test results as console report, data frame, or Excel file.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-HTestReporter-print_console}{\code{HTestReporter$print_console()}}
\item \href{#method-HTestReporter-to_dataframe}{\code{HTestReporter$to_dataframe()}}
\item \href{#method-HTestReporter-export_excel}{\code{HTestReporter$export_excel()}}
\item \href{#method-HTestReporter-clone}{\code{HTestReporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestReporter-print_console"></a>}}
\if{latex}{\out{\hypertarget{method-HTestReporter-print_console}{}}}
\subsection{Method \code{print_console()}}{
Print console report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestReporter$print_console(
  result,
  interpret = TRUE,
  interpreter = NULL,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}

\item{\code{interpret}}{Whether to output interpretation}

\item{\code{interpreter}}{StatInterpreter instance}

\item{\code{audience}}{Audience level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestReporter-to_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-HTestReporter-to_dataframe}{}}}
\subsection{Method \code{to_dataframe()}}{
Convert to data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestReporter$to_dataframe(result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestReporter-export_excel"></a>}}
\if{latex}{\out{\hypertarget{method-HTestReporter-export_excel}{}}}
\subsection{Method \code{export_excel()}}{
Export to Excel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestReporter$export_excel(
  result,
  path = NULL,
  excel_exporter = NULL,
  audience = "manager"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{Test result}

\item{\code{path}}{Output path}

\item{\code{excel_exporter}}{ExcelExporter instance}

\item{\code{audience}}{Audience level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HTestReporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HTestReporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HTestReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
