% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalityPlotter.R
\name{NormalityPlotter}
\alias{NormalityPlotter}
\title{NormalityPlotter: Normality test plotting coordinator}
\description{
Automatically select appropriate visualization based on test results.
All plotting functions reuse iQualityR.plot package functions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{theme_obj}}{Theme object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NormalityPlotter-new}{\code{NormalityPlotter$new()}}
\item \href{#method-NormalityPlotter-plot}{\code{NormalityPlotter$plot()}}
\item \href{#method-NormalityPlotter-plot_hist}{\code{NormalityPlotter$plot_hist()}}
\item \href{#method-NormalityPlotter-plot_qq}{\code{NormalityPlotter$plot_qq()}}
\item \href{#method-NormalityPlotter-plot_pp}{\code{NormalityPlotter$plot_pp()}}
\item \href{#method-NormalityPlotter-plot_combined}{\code{NormalityPlotter$plot_combined()}}
\item \href{#method-NormalityPlotter-clone}{\code{NormalityPlotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-new"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-new}{}}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$new(theme = "academic")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theme}}{Theme name or IqrTheme object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-plot"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-plot}{}}}
\subsection{Method \code{plot()}}{
Automatically plot based on test results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$plot(
  x,
  result = NULL,
  plot_type = "auto",
  add_confidence = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Sample data}

\item{\code{result}}{Test result list (NormalityAnalyzer output)}

\item{\code{plot_type}}{Plot type ("auto", "hist", "qq", "pp", "combined")}

\item{\code{add_confidence}}{Whether to add confidence band for QQ plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 or patchwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-plot_hist"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-plot_hist}{}}}
\subsection{Method \code{plot_hist()}}{
Plot histogram with normal curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$plot_hist(x, bins = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Sample data}

\item{\code{bins}}{Number of histogram bins}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-plot_qq"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-plot_qq}{}}}
\subsection{Method \code{plot_qq()}}{
Plot QQ diagram
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$plot_qq(x, add_confidence = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Sample data}

\item{\code{add_confidence}}{Whether to add confidence band}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-plot_pp"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-plot_pp}{}}}
\subsection{Method \code{plot_pp()}}{
Plot PP diagram
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$plot_pp(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Sample data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot2 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-plot_combined"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-plot_combined}{}}}
\subsection{Method \code{plot_combined()}}{
Plot combined figure
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$plot_combined(x, add_confidence = FALSE, layout = "1x3")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Sample data}

\item{\code{add_confidence}}{Whether to add confidence band for QQ plot}

\item{\code{layout}}{Layout}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
patchwork object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalityPlotter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NormalityPlotter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalityPlotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
