% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_registry.R
\name{register_dist}
\alias{register_dist}
\title{Register New Distribution}
\usage{
register_dist(
  dist_type,
  d_func,
  p_func,
  q_func,
  defaults,
  validate = NULL,
  is_discrete = FALSE,
  support = "Custom",
  description = "User-defined distribution"
)
}
\arguments{
\item{dist_type}{Distribution type name}

\item{d_func}{Probability density/mass function}

\item{p_func}{Cumulative distribution function}

\item{q_func}{Quantile function}

\item{defaults}{Default parameter values list}

\item{validate}{Parameter validation function}

\item{is_discrete}{Whether it is a discrete distribution}

\item{support}{Support description}

\item{description}{Distribution description}
}
\value{
Invisible NULL
}
\description{
Add a new distribution to the registry
}
\examples{
\dontrun{
register_dist("my_dist", d = function(x, p) ..., p = ..., q = ...,
              defaults = list(...), is_discrete = FALSE,
              description = "My custom distribution")
}
}
